o_parser/requirements.txt
amqp==5.1.1
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==3.2.20
django-adminlte3==0.1.6
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.1.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
inflection==0.5.1
kombu==5.3.1
outcome==1.2.0
packaging==23.1
prompt-toolkit==3.0.39
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-telegram-bot==20.4
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
tzdata==2023.3
undetected-chromedriver==3.5.0
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
websockets==11.0.3
wsproto==1.2.0
